Externalisation&Desactivation de Random finalisation n°1 ConsumptionMacro
